[Cleanup] Rename remoteWidgetHook
